pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hasnain", email = "hasnainpanhwar744@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
